+++ test/test_template.js
@@ -376,6 +376,13 @@
     }, Error, 'Failing invalid modifier');
   });
 
+  test('Expansion errors', function() {
+    raises(function() {
+      var data = {'composite_var': ['multiple', 'values']};
+      URITemplate('{composite_var:3}').expand(data);
+    }, Error, 'Failing prefix modifier after composite variable');
+  });
+
   test('noConflict mode', function() {
     var actual_lib = URITemplate; // actual library; after loading, before noConflict()
     var unconflicted = URITemplate.noConflict();
@@ -385,6 +392,20 @@
 
     // restore for other tests
     window.URITemplate = actual_lib;
+  });
+
+  test('Periods in varnames', function() {
+    var template = new URITemplate('{hello.world.var}');
+    var literal = 'replacement'
+    var data = {'hello.world.var': literal};
+    var expansion = template.expand(data);
+    equal(expansion, literal, 'period in varname');
+  });
+
+  test('Invalid literals', function () {
+    raises(function() {
+      URITemplate('invalid.char}acter').parse();
+    }, Error, 'Failing invalid literal');
   });
 
 })();