+++ test/urls.js
@@ -291,6 +291,55 @@ var urls = [{
       punycode: false
     }
   }, {
+    name: 'empty username with non-empty password',
+    url: 'scheme://:password@www.example.org/',
+    _url: 'scheme://:password@www.example.org/',
+    parts: {
+      protocol: 'scheme',
+      username: null,
+      password: 'password',
+      hostname: 'www.example.org',
+      port: null,
+      path: '/',
+      query: null,
+      fragment: null
+    },
+    accessors: {
+      protocol: 'scheme',
+      username: '',
+      password: 'password',
+      port: '',
+      path: '/',
+      query: '',
+      fragment: '',
+      resource: '/',
+      authority: ':password@www.example.org',
+      origin: 'scheme://:password@www.example.org',
+      userinfo: ':password',
+      subdomain: 'www',
+      domain: 'example.org',
+      tld: 'org',
+      directory: '/',
+      filename: '',
+      suffix: '',
+      hash: '',
+      search: '',
+      host: 'www.example.org',
+      hostname: 'www.example.org'
+    },
+    is: {
+      urn: false,
+      url: true,
+      relative: false,
+      name: true,
+      sld: false,
+      ip: false,
+      ip4: false,
+      ip6: false,
+      idn: false,
+      punycode: false
+    }
+  }, {
     name: 'malformed email in userinfo',
     url: 'scheme://john@doe.com:pass:word@www.example.org/',
     _url: 'scheme://john%40doe.com:pass%3Aword@www.example.org/',